Crate crust_ecs

an ecs system implemented for AQA NEA coursework 2024.

Provides method such as reading and writing arbitrary typed components, editing component in systems and arbitrary functions as systems just to name a few, it’s heavily inspired by Bevy and Amethyst.

To get started, read the Doc of World which should tell you both where should you start as well as a introduction of all of the functionality that this library provides regarding the world.

Modules

component

entity

system

util

world