Trait crust_ecs::component::ComponentSet

pub trait ComponentSet: Send + Sync + 'static { }

An arbitrary tuple of components usable up to and including 12 members, more than that, you have to manually add them. Should never be instantiated nor used anywhere else except for in the World where a ComponentSet have been specified.

Object Safety

This trait is not object safe.