Trait crust_ecs::system::query::WorldQuery

pub trait WorldQuery: Send + Sync + Sized + SuperWorldQuery {
    type Item<'w>: WorldQuery;
    type TupleAccess;

    // Required methods
    unsafe fn get_column_query<'w>(world: &'w World) -> Self::TupleQuery<'w>;
    fn produce_bitset(world: &World) -> BitSet;
}

A trait for querying components in a world with additional safety and utility methods.

It provides methods to produce a bitset representing the components to be queried, to get a column query for iterating over the components, and to specify the access type for the components being queried.

It’s mostly a marker trait, used to mark any query needed to be put into the Query Struct Currently, only accepts Entity and Components as well as tuple combinations of these as the T part of the trait.

When a component is used, it needs to be a reference and need to specify whether a &(const reference) is needed or a &mut (mutable reference) is needed.

Required Associated Types

type Item<'w>: WorldQuery

type TupleAccess

Required Methods

unsafe fn get_column_query<'w>(world: &'w World) -> Self::TupleQuery<'w>

SAFETY:this function should not be ran unless either single threaded or any form of runtime borrow system is produced

fn produce_bitset(world: &World) -> BitSet

Object Safety

This trait is not object safe.