Trait crust_ecs::system::schedule::IntoSchedule

pub trait IntoSchedule<Marker>: Sized + 'static {
    // Required methods
    fn into_schedule(self) -> ScheduleIntermediate;
    fn system_id(&self) -> SystemID;

    // Provided methods
    fn before<Params: SystemParam>(
        self,
        sys: impl IntoSystem<Params>
    ) -> ScheduleIntermediate { ... }
    fn after<Params: SystemParam>(
        self,
        sys: impl IntoSystem<Params>
    ) -> ScheduleIntermediate { ... }
}

A trait for converting a system or a marker into a schedule intermediate state.

This trait allows a system or a marker to be converted into an intermediate state that can be used to specify dependencies (systems that must be executed before or after the current system). It provides methods to specify these dependencies and to complete the schedule by adding the system to the scheduler.

Required Methods

fn into_schedule(self) -> ScheduleIntermediate

Converts the system or marker into a schedule intermediate state.

fn system_id(&self) -> SystemID

Retrieves the system ID of the system or marker.

Provided Methods

fn before<Params: SystemParam>( self, sys: impl IntoSystem<Params> ) -> ScheduleIntermediate

Specifies that the system should be executed before the specified system.

fn after<Params: SystemParam>( self, sys: impl IntoSystem<Params> ) -> ScheduleIntermediate

Specifies that the system should be executed after the specified system.

Object Safety

This trait is not object safe.