Trait crust_ecs::system::System

pub trait System {
    // Required method
    fn run(&self, world: &World);
}

A wrapper outside the already wrapped FunctionSystem. For detail on the implementation, read here or look for System and Function System in Bevy.

A system in essence is any function which implements explicitly one or more SystemParam only and returns nothing. Currently due to lifetime problem which I encountered, it does not yet support closure.

Example

#[derive(Debug)]
struct Foo {
    a: i32,
}
fn firstSystem(foo: Query<&Foo>) {
    for i in foo {
        println!("{:?}", i);
    }
}
fn main() {
    let mut world = World::new();
    world.insert_column::<Foo>(None);
    world.add_entity((Foo{a:4},));
    world.register_system(firstSystem);
    world.run()
}

Required Methods

fn run(&self, world: &World)